// api/seller/client/models/UpdateComboDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { PrismaJson_VariantComboItems } from './PrismaJson_VariantComboItems';

export type UpdateComboDto = {
  name?: string;
  price?: number;
  compareAtPrice?: number;
  SKU?: string;
  photo?: string;
  items?: PrismaJson_VariantComboItems;
};